new_backend/requirements.txt
fastapi
python-dotenv
openai
PyPDF2
uvicorn